+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- changed setting/getting logger in context to use xlog package [#17](https://github.com/xmidt-org/argus/pull/17)
 
 ## [v0.3.4]
 - handle error case of identifier being too large [#14](https://github.com/xmidt-org/argus/pull/14)